extract out logic for getting paged history to use case; need to move list item and paging into inverted-domain
