+++ data/database/src/commonMain/kotlin/ly/david/musicsearch/data/database/mapper/MapToReleaseForListItem.kt
@@ -2,6 +2,7 @@ package ly.david.musicsearch.data.database.mapper
 
 import ly.david.data.core.release.ReleaseForListItem
 
+// Although SQLDelight generates models for us, their types are based on the function names.
 internal fun mapToReleaseForListItem(
     id: String,
     name: String,
