+++ inverted-domain/build.gradle.kts
@@ -10,6 +10,7 @@ kotlin {
             dependencies {
                 // Although we could move this into domain, let's try keeping them separate, or exposing it as api
                 implementation(projects.data.core)
+                implementation(projects.domain) // TODO: remove
 
                 implementation(compose.runtime)
                 implementation(libs.koin.annotations)
